separate remaining sync service tests
